Release v1.11.0

+++ Middlewares/ST/STM32_WPAN/thread/openthread/stack/include/openthread/platform/logging.h
@@ -110,13 +110,14 @@ extern "C" {
  * This type represents the log level.
  *
  */
-typedef uint8_t otLogLevel;
+typedef int otLogLevel;
 
 /**
  * This enumeration represents log regions.
  *
  */
-typedef enum otLogRegion {
+typedef enum otLogRegion
+{
     OT_LOG_REGION_API      = 1,  ///< OpenThread API
     OT_LOG_REGION_MLE      = 2,  ///< MLE
     OT_LOG_REGION_ARP      = 3,  ///< EID-to-RLOC mapping.
